fixed size for spell:move point
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2313 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/spell/spell.cpp
@@ -211,7 +211,7 @@ const char* Spell::GetDescription(void) const
     return _(spells[id].description);
 }
 
-u8 Spell::MovePoint(void) const
+u16 Spell::MovePoint(void) const
 {
     return spells[id].mp;
 }
